\name{animation-defunct}
\alias{animation-defunct}
\alias{ani.news}
\alias{highlight.def}
\alias{write.rss}
\title{Defunct functions in the animation package}
\description{
  Some functions which were included in old versions of
  this package but irrelevant to animations.
}
\details{
  \code{highlight.def} is irrelevant to animations and
  hence it is no longer in this package; the source code
  was moved to
  \url{http://yihui.name/en/2007/09/r-language-definition-file-for-highlight/}.

  \code{write.rss} is irrelevant either; the source code
  can be found at
  \url{https://github.com/yihui/animation/raw/1650902e43fe283d4fd0cea8f559e4e5187d60fe/R/write.rss.R}.

  \code{ani.news} is no longer necessary since R has got a
  function \code{\link[utils]{news}} in the \pkg{utils}
  package
}
\author{
  Yihui Xie <\url{http://yihui.name}>
}
\seealso{
  \code{\link[base]{Defunct}}
}
\keyword{internal}
