% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.lowess.R
\name{boot.lowess}
\alias{boot.lowess}
\title{Bootstrapping with LOWESS}
\usage{
boot.lowess(x, y = NULL, f = 2/3, iter = 3, line.col = "#FF000033", ...)
}
\arguments{
\item{x, y, f, iter}{passed to \code{\link{lowess}}}

\item{line.col}{the color of the LOWESS lines}

\item{...}{other arguments passed to the scatterplot by \code{\link{plot}}}
}
\description{
Sample the original data with replacement and fit LOWESS curves accordingly.
}
\details{
We keep on resampling the data and finally we will see several bootstrapped
LOWESS curves, which may give us a rough idea about a ``confidence interval''
of the LOWESS fit.
}
\references{
Examples at \url{https://yihui.name/animation/example/boot-lowess/}
}
\author{
Yihui Xie
}
