\name{iatemp}
\alias{iatemp}
\alias{iatemp}
\title{Average yearly temperatures in central Iowa.}
\usage{data(iatemp)}
\description{Average yearly temperatures in central Iowa.}
\details{The format is:

Time-Series [1:116] from 1895 to 2010: 32.7 27.8 32.7 30.4 42.6 31.9 34.5 39.8 32.6 39.6 ...
Source: \url{http://www.wrcc.dri.edu/cgi-bin/divplot1_form.pl?1305}}
\docType{data}
\keyword{datasets}
\examples{data(iatemp)
plot(iatemp)}

