% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdf.R
\name{qpdf}
\alias{qpdf}
\title{A wrapper for the PDF toolkit qpdf}
\usage{
qpdf(input, output, options = "--stream-data=compress")
}
\arguments{
\item{input}{the path of the input PDF file}

\item{output}{the path of the output (if missing, \code{output} will be the
same as \code{input})}

\item{options}{options for \command{qpdf} (default to be
\code{'--stream-data=compress'}, i.e. compress the PDF files)}
}
\value{
if \code{ani.options('qpdf')} is non-\code{NULL}, then this function
  returns the status of the operation (\code{0} for success; see
  \code{\link{system}}); otherwise a warning will be issued
}
\description{
If the tool qpdf is available in the system, it will be called to manipulate
the PDF files (especially to compress the PDF files).
}
\details{
This is a wrapper to call \command{qpdf}. The path of \command{qpdf} should
be set via \code{\link{ani.options}(qpdf = 'path/to/qpdf')}.

See the reference for detailed usage of \command{qpdf}.
}
\references{
Examples at \url{https://yihui.org/animation/example/qpdf/}

qpdf official website \url{http://qpdf.sourceforge.net/}
}
\author{
Yihui Xie
}
