\name{ani.start}
\alias{ani.start}
\alias{ani.stop}
\title{Generate an HTML animation page}
\usage{
ani.start(...)

ani.stop(...)
}
\arguments{
  \item{...}{arguments passed to \code{\link{ani.options}}
  to set animation parameters}
}
\description{
These functions are defunct and should not be used any
more.
}
\keyword{internal}

