\docType{data}
\name{iatemp}
\alias{iatemp}
\title{Average yearly temperatures in central Iowa}
\format{Time-Series [1:116] from 1895 to 2010: 32.7 27.8 32.7 30.4 42.6 31.9
  34.5 39.8 32.6 39.6 ...}
\source{
\url{http://www.wrcc.dri.edu/cgi-bin/divplot1_form.pl?1305}
}
\description{
Temperatures in central Iowa over 106 years.
}
\examples{
plot(iatemp)
}

