% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.system.R
\name{sample.system}
\alias{sample.system}
\title{Demonstration for the systematic sampling}
\usage{
sample.system(nrow = 10, ncol = 10, size = 15, p.col = c("blue", "red"), p.cex = c(1, 
    3))
}
\arguments{
\item{nrow}{the desired number of rows of the sample frame.}

\item{ncol}{the desired number of columns of the sample frame.}

\item{size}{the sample size.}

\item{p.col, p.cex}{different colors / magnification rate to annotate the
population and the sample}
}
\value{
None (invisible \code{NULL}).
}
\description{
The whole sample frame is denoted by a matrix (\code{nrow * ncol}) in the
plane, and the sample points with equal intervals are drawn out according to
a random starting point. The points being sampled are marked by red circles.
}
\references{
Examples at \url{https://yihui.name/animation/example/sample-system/}
}
\seealso{
\code{\link{sample}}, \code{\link{sample.simple}},
  \code{\link{sample.cluster}}, \code{\link{sample.ratio}},
  \code{\link{sample.strat}}
}
\author{
Yihui Xie
}
