% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation-package.R
\docType{data}
\name{vanke1127}
\alias{vanke1127}
\title{Stock prices of Vanke Co., Ltd on 2009/11/27}
\format{
A data frame with 2831 observations on the following 2 variables.
  \describe{ \item{time}{POSIXt: the time corresponding to stock prices}
  \item{price}{a numeric vector: stock prices} }
}
\source{
This data can be obtained from most stock websites.
}
\description{
This is a sample of stock prices of the Vanke Co., Ltd on 2009/11/27.
}
