% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi.lilac.chaser.R
\name{vi.lilac.chaser}
\alias{vi.lilac.chaser}
\title{Visual Illusions: Lilac Chaser}
\usage{
vi.lilac.chaser(np = 16, col = "magenta", bg = "gray", p.cex = 7, c.cex = 5)
}
\arguments{
\item{np}{number of points}

\item{col}{color of points}

\item{bg}{background color of the plot}

\item{p.cex}{magnification of points}

\item{c.cex}{magnification of the center cross}
}
\value{
\code{NULL}
}
\description{
Stare at the center cross for a few (say 30) seconds to experience the
phenomena of the illusion.
}
\details{
Just try it out.
}
\note{
In fact, points in the original version of `Lilac Chaser' are
  \emph{blurred}, which is not implemented in this function.
}
\references{
Examples at \url{https://yihui.org/animation/example/vi-lilac.chaser/}
}
\seealso{
\code{\link{points}}
}
\author{
Yihui Xie
}
